Depot + Pipeline documentation updates
