+++ protocol/contracts/farm/facets/FarmFacet.sol
@@ -11,11 +11,12 @@ import {LibEth} from "../../libraries/Token/LibEth.sol";
 import {LibFunction} from "../../libraries/LibFunction.sol";
 
 /**
- * @title Farmfacet - Users call any function in Beanstalk
+ * @title Farm Facet
+ * @notice Perform mutliple Beanstalk function calls in a single transaction
  **/
 
-
-// Advanced Data is a function call that allows for return values from existing functions 
+// Advanced Data is a function call that allows for return values from existing functions
+// See LibFunction.buildAdvancedCalldata for details
 struct AdvancedData {
     bytes callData;
     bytes farmData;
@@ -24,14 +25,11 @@ struct AdvancedData {
 contract FarmFacet {
     AppStorage internal s;
 
-    /*
-     * Farm Function
-     */
-
-
-    /// @notice Call multiple functions in Beanstalk and return the data from all of them if they all succeed
-    /// @param data The encoded function data for each of the calls to make to this contract.
-    /// @return results The results from each of the calls passed in via data
+    /**
+     * @notice Call multiple functions in Beanstalk and return the data from all of them if they all succeed
+     * @param data The encoded function data for each of the calls to make to this contract
+     * @return results The results from each of the calls passed in via data
+    **/
     function farm(bytes[] calldata data)
         external
         payable
@@ -44,9 +42,12 @@ contract FarmFacet {
         }
     }
 
-    /// @notice Call multiple functions in Beanstalk using 
-    /// @param data The encoded function data for each of the calls to make to this contract.
-    /// @return results The results from each of the calls passed in via data
+    /**
+     * @notice Perform a list of advanced function calls on Beanstalk
+     * @param data The encoded function data for each of the calls to make to this contract
+     * See LibFunction.buildAdvancedCalldata for details on advanced data
+     * @return results The results from each of the calls passed in via data
+    **/
     function advancedFarm(AdvancedData[] calldata data)
         external
         payable
@@ -65,8 +66,7 @@ contract FarmFacet {
     {
         bytes1 pipeType = d.farmData[0];
         // 0x00 -> Normal pipe: Standard function call
-        // 0x01 -> Inject 1 previous return value into function call calldata
-        // 0x02 -> Inject n previous return values into function call calldata
+        // else > Advanced pipe: Copy return data into function call through buildAdvancedCalldata
         if (pipeType == 0x00) {
             result = _farm(d.callData);
         } else {
@@ -75,33 +75,26 @@ contract FarmFacet {
         }
     }
 
-    // delegatecall a Beanstalk function using calldata data.
+    // delegatecall a Beanstalk function using calldata data
     function _farm(bytes calldata data) private returns (bytes memory result) {
-        bytes4 selector;
-        bool success;
-        assembly {
-            selector := calldataload(data.offset)
-        }
+        bytes4 selector; bool success;
+        assembly { selector := calldataload(data.offset) }
         address facet = LibFunction.facetForSelector(selector);
         (success, result) = facet.delegatecall(data);
         LibFunction.checkReturn(success, result);
     }
 
-    // delegatecall a Beanstalk function using memory data.
+    // delegatecall a Beanstalk function using memory data
     function _farmMem(bytes memory data) private returns (bytes memory result) {
-        bytes4 selector;
-        bool success;
-        assembly {
-            selector := mload(add(data, 32))
-        }
+        bytes4 selector; bool success;
+        assembly { selector := mload(add(data, 32)) }
         address facet = LibFunction.facetForSelector(selector);
         (success, result) = facet.delegatecall(data);
         LibFunction.checkReturn(success, result);
     }
 
-    // With Eth signals to Beanstalk functions that 
-    // they should not refund Eth at the end of the function 
-    // because the function is wrapped in a Farm function.
+    // signals to Beanstalk functions that they should not refund Eth 
+    // at the end of the function because the function is wrapped in a Farm function
     modifier withEth() {
         if (msg.value > 0) s.isFarm = 2;
         _;
