+++ protocol/scripts/deploy.js
@@ -108,7 +108,7 @@ async function main(scriptName, verbose = true, mock = false, reset = true) {
     marketplaceFacet,
     ownershipFacet,
     pauseFacet,
-    pipelineFacet,
+    depotFacet,
     seasonFacet,
     siloFacet,
     fertilizerFacet,
@@ -125,7 +125,7 @@ async function main(scriptName, verbose = true, mock = false, reset = true) {
       'MockFundraiserFacet',
       'MockMarketplaceFacet',
       'PauseFacet',
-      'PipelineFacet',
+      'DepotFacet',
       'MockSeasonFacet',
       'MockSiloFacet',
       'MockFertilizerFacet',
@@ -144,7 +144,7 @@ async function main(scriptName, verbose = true, mock = false, reset = true) {
       'MarketplaceFacet',
       'OwnershipFacet',
       'PauseFacet',
-      'PipelineFacet',
+      'DepotFacet',
       'SeasonFacet',
       'SiloFacet',
       'FertilizerFacet',
@@ -181,7 +181,7 @@ async function main(scriptName, verbose = true, mock = false, reset = true) {
       ['MarketplaceFacet', marketplaceFacet],
       ['OwnershipFacet', ownershipFacet],
       ['PauseFacet', pauseFacet],
-      ['PipelineFacet', pipelineFacet],
+      ['DepotFacet', depotFacet],
       ['SeasonFacet', seasonFacet],
       ['SiloFacet', siloFacet],
       ['FertilizerFacet', fertilizerFacet],
@@ -224,7 +224,7 @@ async function main(scriptName, verbose = true, mock = false, reset = true) {
     marketplaceFacet,
     ownershipFacet,
     pauseFacet,
-    pipelineFacet,
+    depotFacet,
     seasonFacet,
     siloFacet,
     fertilizerFacet,
