+++ protocol/utils/helpers.js
@@ -1,0 +1,5 @@
+function toBN(a) {
+return ethers.BigNumber.from(a)
+}
+
+exports.toBN = toBN
