+++ protocol/utils/index.js
@@ -3,11 +3,8 @@ const { impersonateSigner, impersonateBeanstalkOwner } = require("./signer.js");
 const { mintUsdc, mintBeans, mintEth } = require("./mint.js")
 const { readPrune } = require("./read.js")
 const { signSiloDepositTokenPermit, signSiloDepositTokensPermit } = require("./permit.js");
-
-
-function toBN(a) {
-    return ethers.BigNumber.from(a)
-  }
+const { packAdvanced, encodeAdvancedData, decodeAdvancedData } = require('./function.js')
+const { toBN } = require('./helpers.js')
 
 exports.toBN = toBN
 exports.getBeanstalk = getBeanstalk
@@ -24,3 +21,6 @@ exports.getPrice = getPrice
 exports.readPrune = readPrune
 exports.signSiloDepositTokenPermit = signSiloDepositTokenPermit
 exports.signSiloDepositTokensPermit = signSiloDepositTokensPermit
+exports.packAdvanced = packAdvanced
+exports.encodeAdvancedData = encodeAdvancedData
+exports.decodeAdvancedData = decodeAdvancedData
